MB-100: Fix compiler warnings
Reviewed-on: http://review.couchbase.org/26937
Tested-by: Michael Wiederhold <mike@couchbase.com>

+++ src/scheduler.cc
@@ -275,7 +275,7 @@ void ExecutorPool::unregisterBucket(EventuallyPersistentEngine *engine) {
     bucketRegistry.erase(itr);
     lh.unlock();
 
-    for (int tidx = 0; tidx < threads.size(); ++tidx) {
+    for (size_t tidx = 0; tidx < threads.size(); ++tidx) {
         LOG(EXTENSION_LOG_INFO,
             "Waiting for thread[%d] to finish in bucket: %s", engine->getName());
         threads[tidx]->stop();
